+++ RELEASE-NOTES.md
@@ -24,6 +24,8 @@
 ## Fixes and improvements
 * The `snow app run` command now allows upgrading to unversioned mode from a versioned or release mode application installation
 * The `snow app teardown` command now allows dropping a package with versions when the `--force` flag is provided
+* Added support for user stages in stage execute command
+
 
 # v2.6.0
 ## Backward incompatibility